Create a separate cache, move TokensError, and clean up
