+++ chainstate/test-suite/src/tests/fungible_tokens.rs
@@ -13,10 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use chainstate::{BlockError, BlockSource, ChainstateError};
+use chainstate::{BlockError, BlockSource, ChainstateError, TokensError};
 use chainstate::{CheckBlockError, CheckBlockTransactionsError, ConnectTransactionError};
 use chainstate_test_framework::{TestBlockInfo, TestFramework, TransactionBuilder};
-use common::chain::tokens::TokensError;
 use common::primitives::{id, Id};
 use common::{
     chain::{
