+++ chainstate/types/src/error.rs
@@ -16,7 +16,7 @@
 use thiserror::Error;
 
 use common::{
-    chain::{tokens::TokensError, Block, GenBlock},
+    chain::{Block, GenBlock},
     primitives::{BlockHeight, Id},
 };
 
@@ -43,6 +43,6 @@ pub enum PropertyQueryError {
     },
     #[error("Genesis block has no header")]
     GenesisHeaderRequested,
-    #[error("Tokens error: {0}")]
-    TokensError(TokensError),
+    // #[error("Tokens error: {0}")]
+    // TokensError(TokensError),
 }
